+++ src/app/colorpicker/colorpicker.demo.ts
@@ -88,6 +88,5 @@ export class ColorPickerDemoComponent implements OnInit {
       color: '',
       label: ''
     };
-
   }
 }